treewide: move header files into dedicated directory
