+++ lib/fs.c
@@ -26,7 +26,7 @@
 #include <pwd.h>
 #include <glob.h>
 
-#include "../module.h"
+#include "ucode/module.h"
 
 #define err_return(err) do { last_error = err; return NULL; } while(0)
