+++ source.c
@@ -16,7 +16,7 @@
 
 #include <string.h>
 
-#include "source.h"
+#include "ucode/source.h"
 
 
 uc_source_t *
